requirements.txt
tensorflow=2.2.0
opencv-python=4.5.1
numpy=1.19.2
imutils
python-magic
